Fix for larger screens
I repurposed your code for a project that I'm working on (as I found that it was more optimized than the original from Sam). 

However, I realized that the rendering of both on your netlify instance and my project wasn't great on my iMac 27". 

This is the fix I came up with - it does the job in my project but you might want to QA it here.